Project skeleton refactor
