+++ termgram/config.py
@@ -1,0 +1,7 @@
+APP_VERSION = 0.1
+APP_NAME = 'termgram'
+
+TELEGRAM_API_ID = 117120
+TELEGRAM_API_HASH = '4f3d31306935d48f1c9cc42b75838521'
+
+TIMESTAMP_FORMAT = '%H:%M'
